requirements.txt
# Core runtime
numpy==1.26.4
networkx==3.2
Pillow==10.1.0
matplotlib==3.8.1

# Optional loaders / DXF
ezdxf==1.0.2

# Testing & tooling
pytest==7.4.2
pytest-cov==4.1.0
black==24.3.0
flake8==7.0.0
mypy==1.5.1
isort==5.12.0
pre-commit==3.5.0
scikit-image>=0.20.0
scipy>=1.11.0
